classification/requirements.txt
timm==0.9.2
h5py
pandas
PyYAML
opencv-python-headless==4.9.0.80
# opencv-python
matplotlib
scikit-learn
scipy
tqdm
mlxtend
tensorboardX
wandb
torchsnooper
pandas
scikit-learn
pycm
open_clip_torch
albumentations
python-dotenv